pyproject.toml
[project]
name = "ucgrants"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["databricks-sdk>=0.53.0"]

[project.optional-dependencies]
test = ["pyspark>=3.5,<4.0", "pytest>=8.3.5"]
dev = [
    "databricks-connect>=16.1.5",
    "mypy>=1.15.0",
    "nox>=2025.5.1",
    "pip>=25.1.1",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "ruff>=0.11.10",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unreachable = true
files = ["ucgrants", "tests"]

[tool.ruff.lint]
select = ["F", "E", "W", "I", "UP", "RUF"]
ignore = ["E501"]
